rename src -> lib
