+++ lib/edit_control.rb
