+++ lib/factory.rb
