+++ lib/line.rb
