+++ lib/task_manager.rb
