+++ lib/view.rb
