+++ test/test_task_extract.rb
@@ -1,4 +1,4 @@
-require_relative '../src/task_extract.rb'
+require_relative '../lib/task_extract.rb'
 require 'test/unit'
 
 class TestTaskExtract < Test::Unit::TestCase
@@ -24,7 +24,7 @@ class TestTaskExtract < Test::Unit::TestCase
   end
 
   def test_task_extract_with_file_striction_not_finished
-    assert_equal ['./src/task.rb', './src/view.rb'], (file_striction_extract './task/sample/file_striction/test_task_file_striction_extract')
+    assert_equal ['./lib/task.rb', './lib/view.rb'], (file_striction_extract './task/sample/file_striction/test_task_file_striction_extract')
     assert_equal [], (file_striction_extract './task/sample/file_striction/test_task_file_striction_extract_1')
   end
 end
